requirements.txt
kopf
psycopg2-binary
kubernetes
PyYAML
kubernetes
